feature/theme-emoticon 컴포넌트 재구성
